[naga] Move some helper functions to `type_methods`
